7.26

+++ src/main/java/com/example/workproject/Service/OperatorService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface OperatorService {
     public String getAllOperator();
-    public String getOperatorByInstitutionDM(int institutionId, int page, int limit);
+    public String getOperatorByInstitutionDM(long institutionId, int page, int limit);
     public void insertOperator(String param[]);
     public void updateOperator(String czydm, String zt);
 }
